+++ lib/src/tokens.dart
@@ -36,7 +36,8 @@ class CommandToken implements Token {
   @override
   bool operator ==(other) {
     if (other is CommandToken) {
-      return this.type == other.type && this.coordinateType == other.coordinateType;
+      return this.type == other.type &&
+          this.coordinateType == other.coordinateType;
     }
     return false;
   }